Move open app (and reduce a dependency)
