+++ keybase/platform_darwin_test.go
@@ -46,7 +46,7 @@ func TestRestart(t *testing.T) {
 	ctx := newContext(&testConfigDarwin{}, testLog)
 	appPath := ctx.config.destinationPath()
 
-	err := process.OpenAppDarwin(appPath, testLog)
+	err := OpenAppDarwin(appPath, testLog)
 	defer func() {
 		process.TerminateAll(process.NewMatcher(appPath, process.PathPrefix, testLog), 200*time.Millisecond, testLog)
 	}()
@@ -56,4 +56,28 @@ func TestRestart(t *testing.T) {
 	// test the error that the app was ok, but the services didn't restart.
 	err = ctx.restart(20*time.Millisecond, 20*time.Millisecond)
 	assert.EqualError(t, err, "There were multiple errors: No process found for Test.app/Contents/SharedSupport/bin/keybase; No process found for Test.app/Contents/SharedSupport/bin/kbfs")
+}
+
+func TestOpenDarwin(t *testing.T) {
+	appPath := filepath.Join(os.Getenv("GOPATH"), "src/github.com/keybase/go-updater/test/Test.app")
+	matcher := process.NewMatcher(appPath, process.PathEqual, testLog)
+	defer process.TerminateAll(matcher, 200*time.Millisecond, testLog)
+	err := OpenAppDarwin(appPath, testLog)
+	assert.NoError(t, err)
+}
+
+func TestOpenDarwinError(t *testing.T) {
+	binErr := filepath.Join(os.Getenv("GOPATH"), "src/github.com/keybase/go-updater/test/err.sh")
+	appPath := filepath.Join(os.Getenv("GOPATH"), "src/github.com/keybase/go-updater/test/Test.app")
+	err := openAppDarwin(binErr, appPath, time.Millisecond, testLog)
+	assert.Error(t, err)
+}
+
+func TestFindPIDsLaunchd(t *testing.T) {
+	procPath := "/sbin/launchd"
+	matcher := process.NewMatcher(procPath, process.PathEqual, testLog)
+	pids, err := process.FindPIDsWithMatchFn(matcher.Fn(), testLog)
+	assert.NoError(t, err)
+	t.Logf("Pids: %#v", pids)
+	require.True(t, len(pids) >= 1)
 }
